+++ solidity/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @hyperlane-xyz/core
 
+## 7.1.8
+
+### Patch Changes
+
+- ba4deea: Revert workspace dependency syntax.
+- Updated dependencies [ba4deea]
+  - @hyperlane-xyz/utils@13.1.1
+
+## 7.1.7
+
+### Patch Changes
+
+- Updated dependencies [f41f766]
+  - @hyperlane-xyz/utils@13.1.0
+
 ## 7.1.6
 
 ### Patch Changes