+++ typescript/cosmos-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hyperlane-xyz/cosmos-sdk
 
+## 13.1.1
+
+### Patch Changes
+
+- ba4deea: Revert workspace dependency syntax.
+- Updated dependencies [ba4deea]
+  - @hyperlane-xyz/cosmos-types@13.1.1
+
+## 13.1.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@13.1.0
+
 ## 13.0.0
 
 ### Minor Changes